+++ src/pptx2/__init__.py
@@ -35,6 +35,7 @@ from pptx2.opc.constants import CONTENT_TYPE as CT
 from pptx2.opc.package import PartFactory
 from pptx2.parts.chart import ChartPart
 from pptx2.parts.coreprops import CorePropertiesPart
+from pptx2.parts.customprops import CustomPropertiesPart
 from pptx2.parts.diagram import (
     DiagramColorsPart,
     DiagramDataPart,
@@ -98,6 +99,7 @@ content_type_to_part_class_map: dict[str, type[Part]] = {
     CT.PML_TEMPLATE_MAIN: PresentationPart,
     CT.PML_SLIDESHOW_MAIN: PresentationPart,
     CT.OPC_CORE_PROPERTIES: CorePropertiesPart,
+    CT.OFC_CUSTOM_PROPERTIES: CustomPropertiesPart,
     CT.PML_NOTES_MASTER: NotesMasterPart,
     CT.PML_NOTES_SLIDE: NotesSlidePart,
     CT.PML_SLIDE: SlidePart,
@@ -136,6 +138,7 @@ PartFactory.part_type_for.update(content_type_to_part_class_map)
 del (
     ChartPart,
     CorePropertiesPart,
+    CustomPropertiesPart,
     DiagramColorsPart,
     DiagramDataPart,
     DiagramLayoutPart,